+++ login.php
@@ -17,8 +17,6 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
             $errors[$field] = "Это поле надо заполнить";
         }
     }
-
-
     $email = mysqli_real_escape_string($link, $form['email']);
     $user = get_user($link, $email);
     if(!count($errors)) {
@@ -27,12 +25,10 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
         } else {
             $errors['password'] = 'Неверный пароль';
         }
-
         if ($form['email'] !== $user["email"]) {
             $errors['email'] = "Данный email не существует";
         }
     }
-
     if (count($errors)) {
         $page_content = include_template('login.php', ['form' => $form, 'errors' => $errors, 'email' => $email, 'equipments' => $categories]);
     }
